add 2

+++ src/model/Customer.java
@@ -3,7 +3,7 @@ package model;
 public class Customer extends Person {
 
 
-    public Customer(double id, String name, int age) {
+    public Customer(int id, String name, int age) {
         super(id, name, age);
     }
 }
