+++ src/model/Seller.java
@@ -3,7 +3,7 @@ package model;
 import model.Person;
 
 public class Seller extends Person {
-    public Seller(double id, String name, int age) {
+    public Seller(int id, String name, int age) {
         super(id, name, age);
     }
 }
